+++ DeliveryMicroservice/src/main/java/tn/fst/igl5/delivery_microservice/domain/DeliveryPerson.java
@@ -43,9 +43,9 @@ public class DeliveryPerson {
     private String plateNumber;
 
     @Column
-    private Long currentLat;
+    private Double currentLat;
     @Column
-    private Long currentLng;
+    private Double currentLng;
 
     @Column
     private boolean available;
